pyproject.toml
[tool.poetry]
name = "padel-shop"
version = "0.1.0"
description = "Padel Webshop created using Django DRF"
authors = ["grilo13"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
numpy = "^1.23.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
